+++ c_src/sys/syscall_define.c
@@ -19,7 +19,7 @@
 #include <linux/unistd.h>
 #include <elf.h>
 #include <linux/audit.h>
-#include "alcove_syscall.h"
+#include "alcove_syscall_constants.h"
 #endif
 
 /*
